predictor/requirements.txt
audioread==3.0.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
decorator==5.1.1
Flask==3.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
msgpack==1.1.0
numba==0.60.0
numpy==1.26.2
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.14.1
soundfile==0.12.1
soxr==0.5.0.post1
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
